update auth token resp types

+++ strava/strava.go
@@ -84,8 +84,8 @@ type Authorization struct {
 
 type authTokenResp struct {
 	TokenType    string  `json:"token_type"`
-	ExpiresAt    uint32  `json:"expires_at"`
-	ExpiresIn    uint32  `json:"expires_in"`
+	ExpiresAt    uint64  `json:"expires_at"`
+	ExpiresIn    uint64  `json:"expires_in"`
 	RefreshToken string  `json:"refresh_token"`
 	AccessToken  string  `json:"access_token"`
 	Athlete      Athlete `json:"athlete"`
